:recycle: Integrate UserTaskConverter into UserTaskService
