+++ src/converters/index.ts
@@ -1,2 +1,1 @@
-export * from './user_task_converter';
 export * from './manual_task_converter';
